# make Sum preserve the internal state