Fixes #2.

No measurable change in speed.

Keeping the blake2b_state inline inside the digest struct instead of as a pointer would get rid of an allocation bringing it back down to 4 per op. I can follow up with that in another PR if you want (or do it here).

```
benchmark              old ns/op     new ns/op     delta
BenchmarkBlake2B-4     1754047       1727184       -1.53%

benchmark              old MB/s     new MB/s     speedup
BenchmarkBlake2B-4     597.80       607.10       1.02x

benchmark              old allocs     new allocs     delta
BenchmarkBlake2B-4     4              5              +25.00%

benchmark              old bytes     new bytes     delta
BenchmarkBlake2B-4     625           1009          +61.44%
```