@codahale 2015-09-23
If you like, you can add the inlining to this PR.